Display: fix system destoryed the buffer on screen.
PD#163001

buffer fd is not clone and passed to display.
When layer destoryed, the buffer will destoryed also,
but display is reading it which may cause screen distored.

So,
1) Clone the buffer hnd, when we get buffer hnd from surfaceflinger.
So we can keep the buffer hnd, when layer in surfaceflinger destroyed.
2) Dup buffer fd to display, display will close it after display.
So display can also display even layer already destroyed.

Change-Id: I206f32493cb57c149364642b3d959906062906a7