hwc: add vendor composer service [1/3]
PD#SWPL-18099

Problem:
When playing a 4k video and press back in YouTube, SurfaceView will not
be destroyed. And in composer service process, there is a composerLayer.
CompserLayer has a ComposerHandleCache which will hold the graphic
buffer handles cached in it and lead to ion orphaned memory.

Solution:
Add vendor composer service.

Verify:
Verify on newton

Change-Id: Ia7cac783941354ada82639990eadadac38603dc8
Signed-off-by: baocheng sun <baocheng.sun@amlogic.com>