hwc2: fix ui flicking on SMPTE mode [1/1]
PD#SWPL-124188

Problem:
UI is slightly shaken on SMPTE mode

Solution:
when UI layer is composed switch of DEVICE and CLIENT
Disp frame of UI layer sending by framework is (0,1,4096,2159)
But client target is (0,0,4096,2160). Then the dst height is switched
between 2160 and 2158. This is a tmp workaround fix.

Verify:
ohm

Change-Id: Id489a2e80c043a6634970ac02ea6ef127d48f74d
Signed-off-by: baocheng sun <baocheng.sun@amlogic.com>